Restful web service & social authentication
On the basis of original restful web services, the new ones could receive the requests from RESTClient tool and return corresponding responds. However, current version is only a test, which has not really implemented the complete functionalities. Besides, in this commit, part of the social authentication login has been finished.